+++ docs/src/components/ComponentDoc/ComponentDocAccessibility.tsx
@@ -33,6 +33,19 @@ const ComponentDocAccessibility = ({ info }) => {
         </p>
       )}
 
+      {info.behaviors && (
+        <p>
+          Available behaviors:{' '}
+          {info.behaviors.map(behavior => (
+            <React.Fragment key={`${behavior.category}-${behavior.name}`}>
+              <a href={`behaviors/${behavior.category}#${_.kebabCase(behavior.name)}`}>
+                {behavior.displayName}
+              </a>{' '}
+            </React.Fragment>
+          ))}
+        </p>
+      )}
+
       {description && <p style={{ whiteSpace: 'pre-line' }}>{description}</p>}
     </>
   )