+++ docs/src/components/Sidebar/Sidebar.tsx
@@ -225,6 +225,9 @@ class Sidebar extends React.Component<any, any> {
                 <Menu.Item as={NavLink} exact to="/" activeClassName="active">
                   Introduction
                 </Menu.Item>
+                <Menu.Item as={NavLink} exact to="/color-palette" activeClassName="active">
+                  Color Palette
+                </Menu.Item>
                 <Menu.Item as={NavLink} exact to="/shorthand-props" activeClassName="active">
                   Shorthand Props
                 </Menu.Item>
@@ -261,6 +264,14 @@ class Sidebar extends React.Component<any, any> {
                 <Menu.Menu>
                   <Menu.Item as={NavLink} exact to="/prototype-chat-pane" activeClassName="active">
                     Chat Pane
+                  </Menu.Item>
+                  <Menu.Item
+                    as={NavLink}
+                    exact
+                    to="/prototype-chat-message-with-popover"
+                    activeClassName="active"
+                  >
+                    Chat message with popover
                   </Menu.Item>
                   <Menu.Item
                     as={NavLink}