+++ src/components/Layout/Layout.tsx
@@ -1,14 +1,12 @@
 import * as React from 'react'
 import * as PropTypes from 'prop-types'
-import * as cx from 'classnames'
+import cx from 'classnames'
 
-import { UIComponent } from '../../lib'
+import { UIComponent, UIComponentProps, commonPropTypes } from '../../lib'
 import { Extendable } from '../../../types/utils'
 import { ICSSInJSStyle } from '../../themes/types'
-import { UIComponentProps } from '../../lib/commonPropInterfaces'
-import { commonUIComponentPropTypes } from '../../lib/commonPropTypes'
 
-export interface LayoutProps extends UIComponentProps<any, any> {
+export interface LayoutProps extends UIComponentProps {
   debug?: boolean
   renderStartArea?: (params: object) => React.ReactNode
   renderMainArea?: (params: object) => React.ReactNode
@@ -50,7 +48,10 @@ class Layout extends UIComponent<Extendable<LayoutProps>, any> {
   static displayName = 'Layout'
 
   static propTypes = {
-    ...commonUIComponentPropTypes,
+    ...commonPropTypes.createCommon({
+      children: false,
+      content: false,
+    }),
     debug: PropTypes.bool,
 
     renderStartArea: PropTypes.func,