+++ src/themes/teams/componentVariables.ts
@@ -8,12 +8,12 @@ export { default as Button } from './components/Button/buttonVariables'
 export { default as ButtonGroup } from './components/Button/buttonVariables'
 
 export { default as Chat } from './components/Chat/chatVariables'
-
 export { default as ChatItem } from './components/Chat/chatItemVariables'
-
 export { default as ChatMessage } from './components/Chat/chatMessageVariables'
 
 export { default as Divider } from './components/Divider/dividerVariables'
+
+export { default as Dropdown } from './components/Dropdown/dropdownVariables'
 
 export { default as Grid } from './components/Grid/gridVariables'