+++ src/themes/teams/components/Icon/iconStyles.ts
@@ -4,7 +4,7 @@ import { fittedStyle } from '../../../../styles/customCSS'
 import { ComponentSlotStylesInput, ICSSInJSStyle, FontIconSpec } from '../../../types'
 import { ResultOf } from '../../../../../types/utils'
 import { IconXSpacing, IconProps } from '../../../../components/Icon/Icon'
-import { pxToRem } from './../../../../lib'
+import { pxToRem } from '../../utils'
 import { getStyle as getSvgStyle } from './svg'
 
 const sizes = new Map([
@@ -24,23 +24,24 @@ const getDefaultFontIcon = (iconName: string) => {
 }
 
 const getFontStyles = (
-  size: string,
+  size: number,
   iconName: string,
   themeIcon?: ResultOf<FontIconSpec>,
 ): ICSSInJSStyle => {
   const { fontFamily, content } = themeIcon || getDefaultFontIcon(iconName)
+  const sizeInRems = pxToRem(size)
 
   return {
     fontFamily,
-    fontSize: getSize(size),
+    fontSize: sizeInRems,
     lineHeight: 1,
     textAlign: 'center',
 
     '::before': {
       content,
       display: 'block',
-      width: getSize(size),
-      height: getSize(size),
+      width: sizeInRems,
+      height: sizeInRems,
     },
   }
 }
@@ -71,7 +72,19 @@ const getPaddedStyle = (): ICSSInJSStyle => ({
   padding: pxToRem(4),
 })
 
-const getSize = size => pxToRem(sizes.get(size))
+const getIconSize = (size, sizeModifier): number => {
+  if (!sizeModifier) {
+    return sizes.get(size)
+  }
+  const modifiedSizes = {
+    large: {
+      x: 24,
+      xx: 28,
+    },
+  }
+
+  return modifiedSizes[size] && modifiedSizes[size][sizeModifier]
+}
 
 const getIconColor = color => color || 'currentColor'
 
@@ -91,8 +104,7 @@ const iconStyles: ComponentSlotStylesInput<IconProps, any> = {
       speak: 'none',
       verticalAlign: 'middle',
 
-      ...(isFontBased &&
-        getFontStyles(size, name, callable(iconSpec && (iconSpec.icon as FontIconSpec))())),
+      ...(isFontBased && getFontStyles(getIconSize(size, v.sizeModifier), name)),
 
       ...(isFontBased && {
         color: getIconColor(v.color),
@@ -105,7 +117,7 @@ const iconStyles: ComponentSlotStylesInput<IconProps, any> = {
       ...getXSpacingStyles(xSpacing, v.horizontalSpace),
 
       ...((bordered || v.borderColor || circular) &&
-        getBorderedStyles(circular, v.borderColor || v.color || theme.siteVariables.bodyColor)),
+        getBorderedStyles(circular, v.borderColor || getIconColor(v.color))),
     }
   },
 
@@ -128,10 +140,12 @@ const iconStyles: ComponentSlotStylesInput<IconProps, any> = {
   },
 
   svg: ({ props: { size, disabled }, variables: v }): ICSSInJSStyle => {
+    const iconSizeInRems = pxToRem(getIconSize(size, v.sizeModifier))
+
     return {
       display: 'block',
-      width: getSize(size),
-      height: getSize(size),
+      width: iconSizeInRems,
+      height: iconSizeInRems,
       fill: getIconColor(v.color),
 
       ...(disabled && {