+++ src/themes/teams/utils/index.ts
@@ -1,0 +1,5 @@
+import { pxToRem as basePxToRem } from '../../../lib'
+
+const themeFontSizeInPx = 14
+
+export const pxToRem = (sizeInPx: number) => basePxToRem(sizeInPx, themeFontSizeInPx)