+++ test/setup.common.ts
@@ -4,11 +4,8 @@
  */
 import * as enzyme from 'enzyme'
 import * as Adapter from 'enzyme-adapter-react-16'
-import { initKeyboardFocusMock } from './specs/lib/keyboardFocusMock'
 
 enzyme.configure({
   adapter: new Adapter(),
   disableLifecycleMethods: true,
 })
-
-initKeyboardFocusMock()