+++ test/specs/behaviors/behavior-test.tsx
@@ -8,7 +8,6 @@ import {
   basicListBehavior,
   basicListItemBehavior,
   buttonBehavior,
-  buttonGroupBehavior,
   iconBehavior,
   imageBehavior,
   inputBehavior,
@@ -31,6 +30,7 @@ import {
 } from 'src/lib/accessibility'
 import { TestHelper } from './testHelper'
 import definitions from './testDefinitions'
+
 const behaviorMenuItems = require('docs/src/behaviorMenu')
 
 const testHelper = new TestHelper()
@@ -39,7 +39,6 @@ testHelper.addTests(definitions)
 testHelper.addBehavior('basicListBehavior', basicListBehavior)
 testHelper.addBehavior('basicListItemBehavior', basicListItemBehavior)
 testHelper.addBehavior('buttonBehavior', buttonBehavior)
-testHelper.addBehavior('buttonGroupBehavior', buttonGroupBehavior)
 testHelper.addBehavior('iconBehavior', iconBehavior)
 testHelper.addBehavior('inputBehavior', inputBehavior)
 testHelper.addBehavior('imageBehavior', imageBehavior)