+++ test/specs/components/Popup/Popup-test.tsx
@@ -1,4 +1,3 @@
-import * as _ from 'lodash'
 import computePopupPlacement, { Position, Alignment } from 'src/components/Popup/positioningHelper'
 import { Placement } from 'popper.js'